minimizes AbstractAssociativeArray to reduce duplication still have AbstractIndexed, AbstractKeyValuePair, and AbstractList to go
