+++ src/Abstract/AbstractAssociativeArray.php
@@ -10,23 +10,15 @@ use PHPAlchemist\Exception\InvalidKeyTypeException;
 use PHPAlchemist\Exception\ReadOnlyDataException;
 use PHPAlchemist\Exception\UnmatchedClassException;
 use PHPAlchemist\Exception\UnmatchedVersionException;
-use PHPAlchemist\Trait\Array\OnClearTrait;
-use PHPAlchemist\Trait\Array\OnInsertTrait;
-use PHPAlchemist\Trait\Array\OnRemoveTrait;
-use PHPAlchemist\Trait\Array\OnSetTrait;
 use PHPAlchemist\Trait\ArrayTrait;
 use PHPAlchemist\Type\Twine;
 
 /**
  * Abstract class for Associative Array (Objectified Array Class).
  */
-abstract class AbstractAssociativeArray implements AssociativeArrayInterface
+abstract class AbstractAssociativeArray extends NaturalArray implements AssociativeArrayInterface
 {
     use ArrayTrait;
-    use OnInsertTrait;
-    use OnRemoveTrait;
-    use OnClearTrait;
-    use OnSetTrait;
 
     public static $serializeVersion = 1;
 
@@ -75,9 +67,9 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
     public function __serialize() : array
     {
         return [
-            'version' => static::$serializeVersion,
-            'model'   => get_class($this),
-            'data'    => $this->data,
+          'version' => static::$serializeVersion,
+          'model'   => get_class($this),
+          'data'    => $this->data,
         ];
     }
 
@@ -97,7 +89,7 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
     /**
      * Add.
      *
-     * @param mixed $key   key to add to array
+     * @param mixed $key key to add to array
      * @param mixed $value value to add to array
      *
      * @return $this
@@ -120,23 +112,6 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
     }
 
     /**
-     * Get a count of the elements of the array.
-     *
-     * @return int
-     */
-    public function count() : int
-    {
-        return count($this->data);
-    }
-
-    public function delete(mixed $key) : void
-    {
-        if (array_key_exists($key, $this->data)) {
-            $this->offsetUnset($key);
-        }
-    }
-
-    /**
      * @param string $glue default: ' '
      *
      * @return StringInterface
@@ -146,60 +121,8 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
         return new Twine(join($glue, $this->data));
     }
 
-    /**
-     * Move back to previous element.
-     *
-     * @return void Any returned value is ignored.
-     */
-    public function prev() : void
-    {
-        $this->position--;
-    }
 
     // region Contractual Obligations
-
-    /**
-     * Whether a offset exists.
-     *
-     * @link   https://php.net/manual/en/arrayaccess.offsetexists.php
-     *
-     * @param mixed $offset <p>
-     *                      An offset to check for.
-     *                      </p>
-     *
-     * @return bool true on success or false on failure.
-     *              </p>
-     *              <p>
-     *              The return value will be casted to boolean if non-boolean was returned.
-     *
-     * @since  5.0.0
-     */
-    public function offsetExists(mixed $offset) : bool
-    {
-        return isset($this->data[$offset]);
-    }
-
-    /**
-     * Offset to retrieve.
-     *
-     * @link   https://php.net/manual/en/arrayaccess.offsetget.php
-     *
-     * @param mixed $offset <p>
-     *                      The offset to retrieve.
-     *                      </p>
-     *
-     * @return mixed Can return all value types.
-     *
-     * @since  5.0.0
-     */
-    public function offsetGet(mixed $offset) : mixed
-    {
-        if ($this->offsetExists($offset)) {
-            return $this->data[$offset];
-        }
-
-        return false;
-    }
 
     /**
      * Offset to set.
@@ -207,7 +130,7 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
      * @link https://php.net/manual/en/arrayaccess.offsetset.php
      *
      * @param mixed $offset The offset to assign the value to.
-     * @param mixed $value  The value to set.
+     * @param mixed $value The value to set.
      *
      * @throws HashTableFullException
      * @throws InvalidKeyTypeException
@@ -220,14 +143,14 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
     public function offsetSet(mixed $offset, mixed $value) : void
     {
         if ($this->isReadOnly()) {
-            throw new ReadOnlyDataException('Invalid call to offsetSet on read-only '.__CLASS__.'.');
+            throw new ReadOnlyDataException('Invalid call to offsetSet on read-only ' . __CLASS__ . '.');
         }
 
         if (!$this->offsetExists($offset)
-            && $this->isFixedSize()
-            && $this->count() == $this->fixedSize
+          && $this->isFixedSize()
+          && $this->count() == $this->fixedSize
         ) {
-            throw new HashTableFullException('Invalid call to offsetSet on '.__CLASS__.'where Size is Fixed and HashTable full.');
+            throw new HashTableFullException('Invalid call to offsetSet on ' . __CLASS__ . 'where Size is Fixed and HashTable full.');
         }
 
         if (!$this->validateKey($offset)) {
@@ -237,8 +160,8 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
         if (isset($this->onInsert) && is_callable($this->onInsert)) {
             $onInsert = $this->onInsert;
             [
-                $offset,
-                $value,
+              $offset,
+              $value,
             ] = $onInsert($offset, $value);
         }
 
@@ -275,87 +198,7 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
             $onRemoveComplete($this->data);
         }
     }
-
-    /**
-     * Return the current element.
-     *
-     * @link https://php.net/manual/en/iterator.current.php
-     *
-     * @return mixed Can return any type.
-     *
-     * @since  5.0.0
-     */
-    public function current() : mixed
-    {
-        return ($this->valid()) ? array_values($this->data)[$this->position] : false;
-    }
-
-    /**
-     * Move forward to next element.
-     *
-     * @link https://php.net/manual/en/iterator.next.php
-     *
-     * @return void Any returned value is ignored.
-     *
-     * @since  5.0.0
-     */
-    public function next() : void
-    {
-        $this->position++;
-    }
-
-    /**
-     * Return the key of the current element.
-     *
-     * @link   https://php.net/manual/en/iterator.key.php
-     *
-     * @return mixed scalar on success, or null on failure.
-     *
-     * @since  5.0.0
-     */
-    public function key() : mixed
-    {
-        return array_keys($this->data)[$this->position];
-    }
-
-    /**
-     * Checks if current position is valid.
-     *
-     * @link   https://php.net/manual/en/iterator.valid.php
-     *
-     * @return bool The return value will be casted to boolean and then evaluated.
-     *              Returns true on success or false on failure.
-     *
-     * @since  5.0.0
-     */
-    public function valid() : bool
-    {
-        return isset(array_values($this->data)[$this->position]);
-    }
-
-    /**
-     * Rewind the Iterator to the first element.
-     *
-     * @link   https://php.net/manual/en/iterator.rewind.php
-     *
-     * @return void Any returned value is ignored.
-     *
-     * @since  5.0.0
-     */
-    public function rewind() : void
-    {
-        $this->position = 0;
-    }
-
     // endRegion
-
-    /**
-     * @return array
-     */
-    public function getData() : array
-    {
-        return $this->data;
-    }
 
     /**
      * @param array $data
@@ -456,38 +299,6 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
     }
 
     /**
-     * Get the value of a specified key and remove from
-     * array.
-     *
-     * @param mixed $key
-     *
-     * @return mixed
-     */
-    public function extract(mixed $key) : mixed
-    {
-        $returnValue = $this->data[$key];
-        $this->delete($key);
-
-        return $returnValue;
-    }
-
-    public function clear() : void
-    {
-        if (isset($this->onClear) && is_callable($this->onClear)) {
-            $onClear = $this->onClear;
-            $onClear($this->data);
-        }
-
-        $this->data = [];
-        $this->rewind();
-
-        if (isset($this->onClearComplete) && is_callable($this->onClearComplete)) {
-            $onClearComplete = $this->onClearComplete;
-            $onClearComplete($this->data);
-        }
-    }
-
-    /**
      * Find the key for $value.
      *
      * @param mixed $value the value to search the array for
@@ -497,10 +308,5 @@ abstract class AbstractAssociativeArray implements AssociativeArrayInterface
     public function search(mixed $value) : mixed
     {
         return array_search($value, $this->data);
-    }
-
-    public function isEmpty() : bool
-    {
-        return empty($this->data);
     }
 }
