+++ tests/Unit/CollectionTest.php
@@ -56,7 +56,7 @@ class CollectionTest extends TestCase
     {
         $serializedObject = 'O:28:"PHPAlchemist\Type\Collection":3:{s:7:"version";i:1;s:5:"model";s:28:"PHPAlchemist\Type\Collection";s:4:"data";a:4:{i:0;s:3:"abc";i:1;s:3:"bcd";i:2;s:3:"cde";i:3;s:3:"def";}}';
         $wrongVersion     = 'O:28:"PHPAlchemist\Type\Collection":3:{s:7:"version";i:3;s:5:"model";s:28:"PHPAlchemist\Type\Collection";s:4:"data";a:4:{i:0;s:3:"abc";i:1;s:3:"bcd";i:2;s:3:"cde";i:3;s:3:"def";}}';
-        $wrongClass       = 'O:28:"PHPAlchemist\Type\Collection":3:{s:7:"version";i:1;s:5:"model";s:28:"PHPAlchemist\Type\Hashtable";s:4:"data";a:4:{i:0;s:3:"abc";i:1;s:3:"bcd";i:2;s:3:"cde";i:3;s:3:"def";}}';
+        $wrongClass       = 'O:28:"PHPAlchemist\Type\Collection":3:{s:7:"version";i:1;s:5:"model";s:27:"PHPAlchemist\Type\Hashtable";s:4:"data";a:4:{i:0;s:3:"abc";i:1;s:3:"bcd";i:2;s:3:"cde";i:3;s:3:"def";}}';
 
         $data = unserialize($serializedObject);
         $this->assertInstanceOf('PHPAlchemist\Type\Collection', $data);
